Update major version of LeGUI library

+++ src/main/java/cc/abro/tow/client/menu/InterfaceStyles.java
@@ -4,7 +4,6 @@ import org.joml.Vector4f;
 import org.liquidengine.legui.style.Background;
 import org.liquidengine.legui.style.Style;
 import org.liquidengine.legui.style.border.SimpleLineBorder;
-import org.liquidengine.legui.style.shadow.Shadow;
 
 public final class InterfaceStyles {
 
@@ -52,12 +51,12 @@ public final class InterfaceStyles {
     public final static int LABEL_HEIGHT_DEBUG = 14;
     public final static int LABEL_HEIGHT_ZERO_RULE = 18;
 
-    public final static int MENU_BUTTON_FONT_SIZE = 30;
-    public final static int BUTTON_FONT_SIZE = 20;
-    public final static int LABEL_FONT_SIZE = BUTTON_FONT_SIZE;
-    public final static int ANALYZER_INFO_FONT_SIZE = 16;
-    public final static int SLIGHTLY_BIG_LABEL_FONT_SIZE = 27;
-    public final static int BIG_LABEL_FONT_SIZE = 35;
+    public final static float MENU_BUTTON_FONT_SIZE = 30;
+    public final static float BUTTON_FONT_SIZE = 20;
+    public final static float LABEL_FONT_SIZE = BUTTON_FONT_SIZE;
+    public final static float ANALYZER_INFO_FONT_SIZE = 16;
+    public final static float SLIGHTLY_BIG_LABEL_FONT_SIZE = 27;
+    public final static float BIG_LABEL_FONT_SIZE = 35;
     public final static int LABEL_ERROR_FONT_SIZE = 18;
 
     public final static Vector4f WHITE_COLOR = new Vector4f(1.0f, 1.0f, 1.0f, 1);
